+++ server/src/main/kotlin/org/kryptonmc/krypton/KryptonServer.kt
@@ -6,10 +6,12 @@ import kotlinx.serialization.hocon.Hocon
 import kotlinx.serialization.hocon.decodeFromConfig
 import net.kyori.adventure.text.Component
 import org.kryptonmc.krypton.api.Server
+import org.kryptonmc.krypton.api.event.events.ticking.TickEndEvent
+import org.kryptonmc.krypton.api.event.events.ticking.TickStartEvent
 import org.kryptonmc.krypton.api.status.StatusInfo
 import org.kryptonmc.krypton.command.KryptonCommandManager
+import org.kryptonmc.krypton.concurrent.DefaultUncaughtExceptionHandler
 import org.kryptonmc.krypton.config.KryptonConfig
-import org.kryptonmc.krypton.console.ConsoleScope
 import org.kryptonmc.krypton.console.ConsoleSender
 import org.kryptonmc.krypton.console.KryptonConsole
 import org.kryptonmc.krypton.encryption.Encryption
@@ -17,6 +19,8 @@ import org.kryptonmc.krypton.entity.entities.KryptonPlayer
 import org.kryptonmc.krypton.event.KryptonEventBus
 import org.kryptonmc.krypton.extension.logger
 import org.kryptonmc.krypton.packet.PacketLoader
+import org.kryptonmc.krypton.packet.out.play.PacketOutTimeUpdate
+import org.kryptonmc.krypton.packet.state.PacketState
 import org.kryptonmc.krypton.plugin.KryptonPluginManager
 import org.kryptonmc.krypton.registry.RegistryManager
 import org.kryptonmc.krypton.registry.tags.TagManager
@@ -31,6 +35,9 @@ import java.nio.file.Files
 import java.nio.file.Path
 import java.security.SecureRandom
 import java.util.*
+import java.util.concurrent.Executors
+import java.util.concurrent.TimeUnit
+import kotlin.system.measureTimeMillis
 
 class KryptonServer : Server {
 
@@ -74,20 +81,37 @@ class KryptonServer : Server {
 
     override lateinit var pluginManager: KryptonPluginManager
 
-    fun start() = runBlocking(Dispatchers.IO) {
+    @Volatile
+    internal var lastTickTime = 0L
+    private var lastOverloadWarning = 0L
+    private var tickCount = 0
+
+    private val tickScheduler = Executors.newSingleThreadScheduledExecutor()
+
+    @Volatile
+    internal var isRunning = true
+
+    fun start() {
         LOGGER.info("Starting Krypton server on ${config.server.ip}:${config.server.port}...")
         val startTime = System.nanoTime()
 
-        ConsoleScope.launch {
+        Thread {
             LOGGER.debug("Starting console handler")
             KryptonConsole(this@KryptonServer).start()
-        }
+        }.apply {
+            name = "Console Handler"
+            uncaughtExceptionHandler = DefaultUncaughtExceptionHandler(logger("CONSOLE"))
+            isDaemon = true
+        }.start()
 
         LOGGER.debug("Loading packets...")
         PacketLoader.loadAll()
 
+        LOGGER.debug("Registering built-in commands...")
+        commandManager.registerBuiltins()
+
         LOGGER.debug("Starting Netty...")
-        val nettyJob = GlobalScope.launch(Dispatchers.IO) {
+        GlobalScope.launch(Dispatchers.IO) {
             nettyProcess.run()
         }
 
@@ -95,7 +119,7 @@ class KryptonServer : Server {
             LOGGER.warn("-----------------------------------------------------------------------------------")
             LOGGER.warn("SERVER IS IN OFFLINE MODE! THIS SERVER WILL MAKE NO ATTEMPTS TO AUTHENTICATE USERS!")
             LOGGER.warn("While this may allow players without full Minecraft accounts to connect, it also allows hackers to connect with any username they choose! Beware!")
-            LOGGER.warn("To get rid of this message, change online_mode to true in config.conf")
+            LOGGER.warn("To get rid of this message, change online-mode to true in config.conf")
             LOGGER.warn("-----------------------------------------------------------------------------------")
         }
 
@@ -107,16 +131,48 @@ class KryptonServer : Server {
             LOGGER.info("Done (${"%.3fs".format(Locale.ROOT, (System.nanoTime() - startTime) / 1.0E9)})! Type \"help\" for help.")
         }
 
+        lastTickTime = System.currentTimeMillis()
+        if (config.server.tickThreshold > 0) WatchdogProcess(this@KryptonServer).start()
+
         Runtime.getRuntime().addShutdownHook(Thread({
             LOGGER.info("Stopping Krypton...")
-            LOGGER.debug("Shutting down scheduler...")
-            scheduler.shutdown()
+            isRunning = false
             LOGGER.info("Shutting down plugins...")
             pluginManager.shutdown()
+            eventBus.unregisterAll()
+
+            // shut down schedulers and disconnect players
+            scheduler.shutdown()
+            tickScheduler.shutdownNow()
+            sessionManager.shutdown()
             LOGGER.info("Goodbye")
         }, "Shutdown Handler"))
 
-        nettyJob.join()
+        tickScheduler.scheduleAtFixedRate({
+            val nextTickTime = System.currentTimeMillis() - lastTickTime
+            if (nextTickTime > 2000L && lastTickTime - lastOverloadWarning >= 15000L) {
+                LOGGER.warn("Woah there! Can't keep up! Running ${nextTickTime}ms (${nextTickTime / 50} ticks) behind!")
+                lastOverloadWarning = lastTickTime
+            }
+            eventBus.call(TickStartEvent(tickCount))
+            val tickTime = measureTimeMillis(::tick)
+            val finishTime = System.currentTimeMillis()
+            eventBus.call(TickEndEvent(tickCount, tickTime, finishTime))
+            lastTickTime = finishTime
+        }, 0, TICK_INTERVAL, TimeUnit.MILLISECONDS)
+    }
+
+    private fun tick() {
+        tickCount++
+        worldManager.worlds.forEach { (_, world) ->
+            if (tickCount % 20 == 0) {
+                val timePacket = PacketOutTimeUpdate(world.time, world.dayTime)
+                sessionManager.sessions.asSequence()
+                    .filter { it.currentState == PacketState.PLAY }
+                    .forEach { it.sendPacket(timePacket) }
+            }
+            world.tick()
+        }
     }
 
     override fun broadcast(message: Component, permission: String?) {
@@ -154,8 +210,8 @@ class KryptonServer : Server {
 
     companion object {
 
+        private const val TICK_INTERVAL = 1000L / 20L // milliseconds in a tick
         private val HOCON = Hocon {}
-
         private val LOGGER = logger<KryptonServer>()
     }
 }