+++ server/src/main/kotlin/org/kryptonmc/krypton/entity/entities/KryptonPlayer.kt
@@ -91,4 +91,6 @@ class KryptonPlayer(
     override fun resetTitle() {
         session.sendPacket(PacketOutTitle(TitleAction.RESET))
     }
+
+    override fun identity() = Identity.identity(uuid)
 }