+++ server/src/main/kotlin/org/kryptonmc/krypton/world/KryptonWorldBorder.kt
@@ -1,9 +1,11 @@
 package org.kryptonmc.krypton.world
 
 import org.kryptonmc.krypton.api.world.Location
+import org.kryptonmc.krypton.api.world.World
 import org.kryptonmc.krypton.api.world.WorldBorder
 
 data class KryptonWorldBorder(
+    override val world: World,
     override val size: Double,
     override val center: Location,
     override val damageMultiplier: Double,